le probleme persite version 2
