--- src/Controller/SecurityController.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace App\Controller;
-use App\Entity\Utilisateurs;
-Use App\Form\RegistrationType;
-use Doctrine\Common\Persistence\ObjectManager;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
-
-class SecurityController extends AbstractController
-{
-    /**
-     * @Route("/inscription", name="security_registration")
-     */
-    public function registration(Request $request, ObjectManager $manager)
-    {
-        $user = new Utilisateurs();
-        $form = $this-> createForm(RegistrationType::class, $user);
-            
-        $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid())
-        {
-            $manager->persist($user);
-            $manager->flush();
-        }
-
-        return $this->render('security/registration.html.twig',[
-            'form'=> $form->createView()
-        ]);
-    }
-}
