+++ src/Repository/UserRepository.php
@@ -2,25 +2,25 @@
 
 namespace App\Repository;
 
-use App\Entity\Utilisateurs;
+use App\Entity\User;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Common\Persistence\ManagerRegistry;
 
 /**
- * @method Utilisateurs|null find($id, $lockMode = null, $lockVersion = null)
- * @method Utilisateurs|null findOneBy(array $criteria, array $orderBy = null)
- * @method Utilisateurs[]    findAll()
- * @method Utilisateurs[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @method User|null find($id, $lockMode = null, $lockVersion = null)
+ * @method User|null findOneBy(array $criteria, array $orderBy = null)
+ * @method User[]    findAll()
+ * @method User[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class UtilisateursRepository extends ServiceEntityRepository
+class UserRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
-        parent::__construct($registry, Utilisateurs::class);
+        parent::__construct($registry, User::class);
     }
 
     // /**
-    //  * @return Utilisateurs[] Returns an array of Utilisateurs objects
+    //  * @return User[] Returns an array of User objects
     //  */
     /*
     public function findByExampleField($value)
@@ -37,7 +37,7 @@ class UtilisateursRepository extends ServiceEntityRepository
     */
 
     /*
-    public function findOneBySomeField($value): ?Utilisateurs
+    public function findOneBySomeField($value): ?User
     {
         return $this->createQueryBuilder('u')
             ->andWhere('u.exampleField = :val')
