+++ templates/ofapp/index.html.twig
@@ -1,0 +1,4 @@
+{% extends "base.html.twig" %}
+{% block body %}
+    <h1>Page d'index</h1>
+{% endblock %}
