+++ templates/security/registration.html.twig
@@ -2,11 +2,11 @@
 {% block body %}
     <h1>Inscription sur le site</h1>
     {{form_start(form)}}
-    {{form_row(form.email,{'label':'Nom d\'utilisateur','attr':{'placeholder': 'Adresse email ...'}})}}
-    {{form_row(form.password,{'label':'Mot de passe','attr':{'placeholder':'Mot de passe ...'}})}}
-    {{form_row(form.confirm_password,{'label':'Confirmation mot de passe','attr':{'placeholder':'Confirmer mot de passe ...'}})}}
-  
+    {{form_row(form.username, {'label':'Nom d\'utilisateur','attr':{'placeholder':'Votre nom d\'utilisateur...'}})}}
+    {{form_row(form.email, {'label':'Adresse Email','attr':{'placeholder':'Votre Adresse email...'}})}}
+    {{form_row(form.password, {'label':'Mot de Passe','attr':{'placeholder':'Votre mot de passe'}})}}
+    {{form_row(form.confirm_password, {'label':'Confirmation','attr':{'placeholder':'Confirmer votre mot de passe ...'}})}}
+
     <button type="submit" class="btn btn-success">Inscription</button>
     {{form_end(form)}}
-    
 {% endblock %}
